Use StorageFactorySwitch and move KafkaStorage creation into `local-kafka` module.
