+++ api-java/build.gradle
@@ -20,7 +20,6 @@
 
 dependencies {
     compile project(path: ':model')
-    compile project(path: ':kafka-storage')
 
     testCompile project(path: ':testutil-api')
 }
