+++ api-java/src/main/java/io/spine/examples/todolist/context/BoundedContexts.java
@@ -22,6 +22,7 @@ package io.spine.examples.todolist.context;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
+import com.google.common.base.Supplier;
 import io.spine.examples.todolist.repository.DraftTasksViewRepository;
 import io.spine.examples.todolist.repository.LabelAggregateRepository;
 import io.spine.examples.todolist.repository.LabelledTasksViewRepository;
@@ -32,17 +33,11 @@ import io.spine.server.BoundedContext;
 import io.spine.server.event.EventBus;
 import io.spine.server.event.EventEnricher;
 import io.spine.server.storage.StorageFactory;
-import io.spine.server.storage.kafka.KafkaStorageFactory;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.time.Duration;
-import java.util.Properties;
+import io.spine.server.storage.StorageFactorySwitch;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.server.storage.kafka.Consistency.STRONG;
+import static io.spine.server.storage.StorageFactorySwitch.newInstance;
 import static io.spine.util.Exceptions.newIllegalStateException;
-import static java.time.temporal.ChronoUnit.MILLIS;
 
 /**
  * Utilities for creation the {@link BoundedContext} instances.
@@ -52,23 +47,10 @@ import static java.time.temporal.ChronoUnit.MILLIS;
  */
 public final class BoundedContexts {
 
-    // TODO:2017-09-15:dmytro.dashenkov: Use storage factory switch.
-    private static final String KAFKA_PRODUCER_PROPS_PATH = "config/kafka-producer.properties";
-    private static final String KAFKA_CONSUMER_PROPS_PATH = "config/kafka-consumer.properties";
-    private static final Duration POLL_AWAIT = Duration.of(50, MILLIS);
-
     /** The name of the Bounded Context. */
     private static final String NAME = "TodoListBoundedContext";
-    private static final StorageFactory DEFAULT_STORAGE_FACTORY;
 
-    static {
-        final Properties producerConfig = loadProperties(KAFKA_PRODUCER_PROPS_PATH);
-        final Properties consumerConfig = loadProperties(KAFKA_CONSUMER_PROPS_PATH);
-        DEFAULT_STORAGE_FACTORY = new KafkaStorageFactory(producerConfig,
-                                                          consumerConfig,
-                                                          STRONG,
-                                                          POLL_AWAIT);
-    }
+    private static final StorageFactorySwitch storageFactorySwitch = newInstance(NAME, false);
 
     private BoundedContexts() {
         // Disable instantiation from outside.
@@ -79,13 +61,17 @@ public final class BoundedContexts {
     }
 
     /**
-     * Creates the {@link BoundedContext} instance
-     * using {@code KafkaStorageFactory} for a single tenant.
+     * Creates the {@link BoundedContext} instance using the {@code StorageFactory} from
+     * the {@link StorageFactorySwitch}.
+     *
+     * <p>Use {@link #injectStorageFactory(StorageFactory)} to override the used
+     * {@code StorageFactory} implementation.
      *
      * @return the {@link BoundedContext} instance
      */
     public static BoundedContext create() {
-        final BoundedContext result = create(DEFAULT_STORAGE_FACTORY);
+        final StorageFactory storageFactory = storageFactorySwitch.get();
+        final BoundedContext result = create(storageFactory);
         return result;
     }
 
@@ -137,6 +123,21 @@ public final class BoundedContexts {
         return eventBus;
     }
 
+    /**
+     * Injects the given {@link StorageFactory} implementation.
+     *
+     * <p>After calling this method, invocation of {@link #create()} creates create a
+     * {@code BoundedContext} with the passed {@code StorageFactory}.
+     *
+     * @param storageFactory the default {@code StorageFactory} to create the BoundedContexts for
+     */
+    public static void injectStorageFactory(StorageFactory storageFactory) {
+        checkNotNull(storageFactory);
+        @SuppressWarnings("Guava") // Spine Java 7 API.
+        final Supplier<StorageFactory> storageFactorySupplier = () -> storageFactory;
+        storageFactorySwitch.init(storageFactorySupplier, storageFactorySupplier);
+    }
+
     @SuppressWarnings("Guava" /* Spine API is Java 7-based
                                  and uses `Optional` from Google Guava. */)
     @VisibleForTesting
@@ -145,31 +146,10 @@ public final class BoundedContexts {
         if (!storageFactory.isPresent()) {
             throw newIllegalStateException("EventBus does not specify a StorageFactory.");
         }
-
         return BoundedContext.newBuilder()
                              .setStorageFactorySupplier(storageFactory::get)
                              .setName(NAME)
                              .setEventBus(eventBus)
                              .build();
-    }
-
-    /**
-     * Loads {@code .properties} file from the classpath by the given filename.
-     *
-     * <p>If the file is not found, an {@link NullPointerException} is thrown.
-     */
-    private static Properties loadProperties(String filename) {
-        final ClassLoader loader = BoundedContexts.class.getClassLoader();
-        final InputStream rawProperties = loader.getResourceAsStream(filename);
-        checkNotNull(rawProperties, "Could not load properties file %s from classpath.", filename);
-
-        final Properties result = new Properties();
-
-        try (InputStream props = rawProperties) {
-            result.load(props);
-        } catch (IOException e) {
-            throw new IllegalStateException(e);
-        }
-        return result;
     }
 }
