+++ deployment/local-kafka/src/main/resources/config/kafka-producer.properties
